Rebuild theme after accidential downgrade
A change committed from an outdated theme resource directory issued
a downgrade of several theme compontents resulting in a old icon font
version, unminimized/unoptimized images and a javascript concat
missing a library dependency.

In order to avoid outdated resource directories we need to come up
with a way to notify the user of dirty states.